backend/requirements.txt
flask
flask-cors
python-dotenv
pandas
numpy
scikit-learn
joblib
numpy
